+++ hpvsim/parameters.py
@@ -84,7 +84,7 @@ def make_pars(**kwargs):
     pars['eff_condoms']         = 0.7   # The efficacy of condoms; https://www.nejm.org/doi/10.1056/NEJMoa053284?url_ver=Z39.88-2003&rfr_id=ori:rid:crossref.org&rfr_dat=cr_pub%20%200www.ncbi.nlm.nih.gov
 
     # Parameters for disease progression
-    pars['clinical_cutoffs']    = {'cin1': 0.33, 'cin2':0.67, 'cin3':0.99} # Parameters the control the clinical cliassification of dysplasia
+    pars['clinical_cutoffs']    = {'cin1': 0.33, 'cin2':0.67} # Parameters the control the clinical cliassification of dysplasia
     pars['hpv_control_prob']    = 0.0 # Probability that HPV is controlled latently vs. cleared
     pars['hpv_reactivation']    = 0.025 # Placeholder
     pars['dur_cancer']          = dict(dist='lognormal', par1=12.0, par2=3.0)  # Duration of untreated invasive cerival cancer before death (years)
@@ -355,6 +355,7 @@ def get_genotype_pars(default=False, genotype=None):
     pars.hpv16.prog_rate    = 0.5 # Rate of progression of dysplasia once it is established. This parameter is used as the growth rate within a logistic function that maps durations to progression probabilities
     pars.hpv16.prog_rate_sd = 0.05 # Standard deviation of the progression rate
     pars.hpv16.rel_beta     = 1  # Baseline relative transmissibility, other genotypes are relative to this
+    pars.hpv16.cancer_prob  = 0.25 # Share of CIN3s that will go on to cancer
     pars.hpv16.imm_boost    = 1.0 # TODO: look for data
     pars.hpv16.sero_prob    = 0.65 # https://www.sciencedirect.com/science/article/pii/S2666679022000027#fig1
 
@@ -365,6 +366,7 @@ def get_genotype_pars(default=False, genotype=None):
     pars.hpv18.prog_rate    = 0.5 # Rate of progression of dysplasia once it is established. This parameter is used as the growth rate within a logistic function that maps durations to progression probabilities
     pars.hpv18.prog_rate_sd = 0.05 # Standard deviation of the progression rate
     pars.hpv18.rel_beta     = 0.72  # Relative transmissibility, current estimate from Harvard model calibration of m2f tx
+    pars.hpv18.cancer_prob  = 0.15  # Share of CIN3s that will go on to cancer
     pars.hpv18.imm_boost    = 1.0 # TODO: look for data
     pars.hpv18.sero_prob    = 0.6 # https://www.sciencedirect.com/science/article/pii/S2666679022000027#fig1
 
@@ -375,6 +377,7 @@ def get_genotype_pars(default=False, genotype=None):
     pars.hpv31.prog_rate    = 0.5 # Rate of progression of dysplasia once it is established. This parameter is used as the growth rate within a logistic function that maps durations to progression probabilities
     pars.hpv31.prog_rate_sd = 0.05 # Standard deviation of the progression rate
     pars.hpv31.rel_beta     = 0.94 # Relative transmissibility, current estimate from Harvard model calibration of m2f tx
+    pars.hpv31.cancer_prob  = 0.05  # Share of CIN3s that will go on to cancer
     pars.hpv31.imm_boost    = 1.0 # TODO: look for data
     pars.hpv31.sero_prob    = 0.53 # https://www.sciencedirect.com/science/article/pii/S2666679022000027#fig1
 
@@ -385,6 +388,7 @@ def get_genotype_pars(default=False, genotype=None):
     pars.hpv33.prog_rate    = 0.5 # Rate of progression of dysplasia once it is established. This parameter is used as the growth rate within a logistic function that maps durations to progression probabilities
     pars.hpv33.prog_rate_sd = 0.05 # Standard deviation of the progression rate
     pars.hpv33.rel_beta     = 0.26 # Relative transmissibility, current estimate from Harvard model calibration of m2f tx
+    pars.hpv33.cancer_prob  = 0.05  # Share of CIN3s that will go on to cancer
     pars.hpv33.imm_boost    = 1.0 # TODO: look for data
     pars.hpv33.sero_prob    = 0.6  # https://www.sciencedirect.com/science/article/pii/S2666679022000027#fig1
 
@@ -395,6 +399,7 @@ def get_genotype_pars(default=False, genotype=None):
     pars.hpv35.prog_rate    = 0.5 # Rate of progression of dysplasia once it is established. This parameter is used as the growth rate within a logistic function that maps durations to progression probabilities
     pars.hpv35.prog_rate_sd = 0.05 # Standard deviation of the progression rate
     pars.hpv35.rel_beta     = 0.5 # Relative transmissibility, current estimate from Harvard model calibration of m2f tx
+    pars.hpv35.cancer_prob  = 0.05  # Share of CIN3s that will go on to cancer
     pars.hpv35.imm_boost    = 1.0 # TODO: look for data
     pars.hpv35.sero_prob    = 0.6  # Assumption, not provided in https://www.sciencedirect.com/science/article/pii/S2666679022000027#fig1
 
@@ -405,6 +410,7 @@ def get_genotype_pars(default=False, genotype=None):
     pars.hpv45.prog_rate    = 0.5 # Rate of progression of dysplasia once it is established. This parameter is used as the growth rate within a logistic function that maps durations to progression probabilities
     pars.hpv45.prog_rate_sd = 0.05 # Standard deviation of the progression rate
     pars.hpv45.rel_beta     = 0.77 # Relative transmissibility, current estimate from Harvard model calibration of m2f tx
+    pars.hpv45.cancer_prob  = 0.15  # Share of CIN3s that will go on to cancer
     pars.hpv45.imm_boost    = 1.0 # TODO: look for data
     pars.hpv45.sero_prob    = 0.25  # https://www.sciencedirect.com/science/article/pii/S2666679022000027#fig1
 
@@ -415,6 +421,7 @@ def get_genotype_pars(default=False, genotype=None):
     pars.hpv51.prog_rate    = 0.5 # Rate of progression of dysplasia once it is established. This parameter is used as the growth rate within a logistic function that maps durations to progression probabilities
     pars.hpv51.prog_rate_sd = 0.05 # Standard deviation of the progression rate
     pars.hpv51.rel_beta     = 0.5 # Relative transmissibility, current estimate from Harvard model calibration of m2f tx
+    pars.hpv51.cancer_prob  = 0.05  # Share of CIN3s that will go on to cancer
     pars.hpv51.imm_boost    = 1.0 # TODO: look for data
     pars.hpv51.sero_prob    = 0.6  # Assumption, not provided in https://www.sciencedirect.com/science/article/pii/S2666679022000027#fig1
 
@@ -425,6 +432,7 @@ def get_genotype_pars(default=False, genotype=None):
     pars.hpv52.prog_rate    = 0.5 # Rate of progression of dysplasia once it is established. This parameter is used as the growth rate within a logistic function that maps durations to progression probabilities
     pars.hpv52.prog_rate_sd = 0.05 # Standard deviation of the progression rate
     pars.hpv52.rel_beta     = 0.623 # Relative transmissibility, current estimate from Harvard model calibration of m2f tx
+    pars.hpv52.cancer_prob  = 0.05  # Share of CIN3s that will go on to cancer
     pars.hpv52.imm_boost    = 1.0 # TODO: look for data
     pars.hpv52.sero_prob    = 0.5  # https://www.sciencedirect.com/science/article/pii/S2666679022000027#fig1
 
@@ -435,6 +443,7 @@ def get_genotype_pars(default=False, genotype=None):
     pars.hpv56.prog_rate    = 0.5 # Rate of progression of dysplasia once it is established. This parameter is used as the growth rate within a logistic function that maps durations to progression probabilities
     pars.hpv56.prog_rate_sd = 0.05 # Standard deviation of the progression rate
     pars.hpv56.rel_beta     = 0.6 # Relative transmissibility, current estimate from Harvard model calibration of m2f tx
+    pars.hpv56.cancer_prob  = 0.15  # Share of CIN3s that will go on to cancer
     pars.hpv56.imm_boost    = 1.0 # TODO: look for data
     pars.hpv56.sero_prob    = 0.6  # Assumption, not provided in https://www.sciencedirect.com/science/article/pii/S2666679022000027#fig1
 
@@ -445,6 +454,7 @@ def get_genotype_pars(default=False, genotype=None):
     pars.hpv58.prog_rate    = 0.5 # Rate of progression of dysplasia once it is established. This parameter is used as the growth rate within a logistic function that maps durations to progression probabilities
     pars.hpv58.prog_rate_sd = 0.05 # Standard deviation of the progression rate
     pars.hpv58.rel_beta     = 0.6 # Relative transmissibility, current estimate from Harvard model calibration of m2f tx
+    pars.hpv58.cancer_prob  = 0.15  # Share of CIN3s that will go on to cancer
     pars.hpv58.imm_boost    = 1.0 # TODO: look for data
     pars.hpv58.sero_prob    = 0.65  # Assumption, not provided in https://www.sciencedirect.com/science/article/pii/S2666679022000027#fig1
 
@@ -455,6 +465,7 @@ def get_genotype_pars(default=False, genotype=None):
     pars.hpv6.prog_rate     = 0.05 # Rate of progression of dysplasia once it is established. This parameter is used as the growth rate within a logistic function that maps durations to progression probabilities
     pars.hpv6.prog_rate_sd  = 0.005 # Standard deviation of the progression rate
     pars.hpv6.rel_beta      = 0.8 # Relative transmissibility, generally calibrated to match available type distributions
+    pars.hpv6.cancer_prob   = 0.0 # Share of CIN3s that will go on to cancer
     pars.hpv6.imm_boost     = 1.0 # TODO: look for data
     pars.hpv6.sero_prob     = 0.6  # Assumption, not provided in https://www.sciencedirect.com/science/article/pii/S2666679022000027#fig1
 
@@ -465,6 +476,7 @@ def get_genotype_pars(default=False, genotype=None):
     pars.hpv11.prog_rate    = 0.05 # Rate of progression of dysplasia once it is established. This parameter is used as the growth rate within a logistic function that maps durations to progression probabilities
     pars.hpv11.prog_rate_sd = 0.005 # Standard deviation of the progression rate
     pars.hpv11.rel_beta     = 0.5 # Relative transmissibility, generally calibrated to match available type distributions
+    pars.hpv11.cancer_prob  = 0.0  # Share of CIN3s that will go on to cancer
     pars.hpv11.imm_boost    = 1.0 # TODO: look for data
     pars.hpv11.sero_prob    = 0.6  # Assumption, not provided in https://www.sciencedirect.com/science/article/pii/S2666679022000027#fig1
 
@@ -475,6 +487,7 @@ def get_genotype_pars(default=False, genotype=None):
     pars.hrhpv.prog_rate    = 0.5 # same value as for all oncogenic types
     pars.hrhpv.prog_rate_sd = 0.05 # same value as for all oncogenic types
     pars.hrhpv.rel_beta     = 0.94 # placeholder, currently assumed to be the same as for 31
+    pars.hrhpv.cancer_prob  = 0.25  # Share of CIN3s that will go on to cancer
     pars.hrhpv.imm_boost    = 1.0 # placeholder, currently assumed to be the same as for 31
     pars.hrhpv.sero_prob    = 0.5 # placeholder, currently assumed to be the same as for 31
 
@@ -485,6 +498,7 @@ def get_genotype_pars(default=False, genotype=None):
     pars.lrhpv.prog_rate    = 0.05 # placeholder, currently assumed to be the same as for 6
     pars.lrhpv.prog_rate_sd = 0.005 # placeholder, currently assumed to be the same as for 6
     pars.lrhpv.rel_beta     = 0.5 # placeholder, currently assumed to be the same as for 6
+    pars.lrhpv.cancer_prob  = 0.0  # Share of CIN3s that will go on to cancer
     pars.lrhpv.imm_boost    = 1.0 # placeholder, currently assumed to be the same as for 6
     pars.lrhpv.sero_prob    = 0.6  # placeholder, currently assumed to be the same as for 6